BackEnd Cleanup, Errors are handled a little bit better,folders added to reflect the purpose of files
