+++ BackEnd/Data/DatabaseMethods/UserMethods.cs
@@ -1,0 +1,147 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using BackEnd.Models;
+using Microsoft.EntityFrameworkCore;
+using BackEnd.Exceptions;
+
+namespace BackEnd.Data
+{
+    public class UserMethods : IUser
+    {
+        protected readonly Context _context;
+        public UserMethods(Context context)
+        {
+            _context = context;
+        }
+
+        public User Authenticate(string username, string password)
+        {
+            if (string.IsNullOrEmpty(username) || string.IsNullOrEmpty(password))
+                return null;
+            var user = _context.Users.SingleOrDefault(x => x.Username == username);
+            if (user == null)
+                return null;
+            if (!VerifyPasswordHash(password, user.PasswordHash, user.PasswordSalt))
+                return null;
+            return user;
+        }
+
+        public User CreateUser(User user, string password)
+        {
+            if (string.IsNullOrWhiteSpace(password))
+                throw new AppException("Password is required");
+            if (_context.Users.Any(x => x.Username == user.Username))
+                throw new AppException($"Username {user.Username} is already taken");
+            byte[] passwordHash, passwordSalt;
+            CreatePasswordHash(password, out passwordHash, out passwordSalt);
+            user.PasswordHash = passwordHash;
+            user.PasswordSalt = passwordSalt;
+            _context.Users.Add(user);
+            _context.SaveChanges();
+            return user;
+        }
+
+        public void Delete(int id)
+        {
+            var user = _context.Users.Find(id);
+            if (user == null)
+                throw new NotFoundException("User not found");
+            if (user != null)
+            {
+                _context.Users.Remove(user);
+                _context.SaveChanges();
+            }
+        }
+
+        public IEnumerable<User> GetAllUsers()
+        {
+            return _context.Users.Include(s => s.Tasks);
+        }
+
+        public User GetUserById(int id)
+        {
+            var user = _context.Users.Include(s => s.Tasks).FirstOrDefault(s => s.Id == id);
+            if (user == null)
+                throw new NotFoundException("User not found");
+            return user;
+        }
+
+        public void Update(User userParam)
+        {
+            var user = _context.Users.Find(userParam.Id);
+            if (user == null)
+                throw new AppException("User not found");
+            // update username if it has changed
+            if (!string.IsNullOrWhiteSpace(userParam.Username) && userParam.Username != user.Username)
+            {
+                // throw error if the new username is already taken
+                if (_context.Users.Any(x => x.Username == userParam.Username))
+                    throw new AppException($"Username {userParam.Username} is already taken");
+                user.Username = userParam.Username.Trim();
+            }
+            // update user properties if provided
+            if (!string.IsNullOrWhiteSpace(userParam.FirstName))
+                user.FirstName = userParam.FirstName;
+
+            if (!string.IsNullOrWhiteSpace(userParam.LastName))
+                user.LastName = userParam.LastName;
+
+            if (userParam.Number != 0)
+                user.Number = userParam.Number;
+
+            if (!string.IsNullOrWhiteSpace(userParam.Country))
+                user.Country = userParam.Country;
+
+            if (!string.IsNullOrWhiteSpace(userParam.City))
+                user.City = userParam.City;
+
+            if (!string.IsNullOrWhiteSpace(userParam.Hobby))
+                user.Hobby = userParam.Hobby;
+
+            _context.Users.Update(user);
+            _context.SaveChanges();
+        }
+
+        public IEnumerable<object> TasksPerUsers(int id)
+        {
+            var user = _context.Users.Find(id);
+            if (user == null)
+                throw new AppException("User not found");
+            var tasks = _context.TasksPerUsers.Where(x => x.UserId == id);
+            return tasks;
+        }
+
+
+        // private helper methods
+        private static void CreatePasswordHash(string password, out byte[] passwordHash, out byte[] passwordSalt)
+        {
+            if (password == null) throw new ArgumentNullException("password");
+            if (string.IsNullOrWhiteSpace(password)) throw new ArgumentException("Value cannot be empty or whitespace only string.", "password");
+            using (var hmac = new System.Security.Cryptography.HMACSHA512())
+            {
+                passwordSalt = hmac.Key;
+                passwordHash = hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(password));
+            }
+        }
+
+        private static bool VerifyPasswordHash(string password, byte[] storedHash, byte[] storedSalt)
+        {
+            if (password == null) throw new ArgumentNullException("password");
+            if (string.IsNullOrWhiteSpace(password)) throw new ArgumentException("Value cannot be empty or whitespace only string.", "password");
+            if (storedHash.Length != 64) throw new ArgumentException("Invalid length of password hash (64 bytes expected).", "passwordHash");
+            if (storedSalt.Length != 128) throw new ArgumentException("Invalid length of password salt (128 bytes expected).", "passwordHash");
+
+            using (var hmac = new System.Security.Cryptography.HMACSHA512(storedSalt))
+            {
+                var computedHash = hmac.ComputeHash(System.Text.Encoding.UTF8.GetBytes(password));
+                for (int i = 0; i < computedHash.Length; i++)
+                {
+                    if (computedHash[i] != storedHash[i]) return false;
+                }
+            }
+            return true;
+        }
+    }
+}
+
