+++ BackEnd/Data/Interface/IProjects.cs
@@ -1,17 +1,10 @@
 using BackEnd.Models;
 using System.Collections.Generic;
 
-namespace BackEnd.Data
+namespace BackEnd.Data.Interface
 {
-    public interface IUser
+    public interface IProjects
     {
-        User Authenticate(string username, string password);
-        IEnumerable<User> GetAllUsers();
-        User GetUserById(int id);
-        User CreateUser(User user, string password);
-        void Update(User user);
-        void Delete(int id);
-        IEnumerable<object> TasksPerUsers(int id);
         IEnumerable<ProjectModel> GetAllProjects();
         ProjectModel GetOneProject(int id);
         IEnumerable<ProjectModel> AddProject(Project project);
