+++ BackEnd/Exceptions/NotFoundException.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace BackEnd.Exceptions
+{
+   public class NotFoundException : Exception
+   {
+       public NotFoundException(string message) : base(message) { }
+   }
+}
