+++ BackEnd/Models/Projects/Project.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Threading.Tasks;
 using System.ComponentModel.DataAnnotations;
 using System.Collections.Generic;
