+++ BackEnd/Models/StatusEnumeration/StatusEnumeration.cs
@@ -1,0 +1,13 @@
+namespace BackEnd.Models.StatusEnumeration
+{
+    public class StatusEnumeration
+    {
+        public enum Status
+        {
+            Pending,
+            Working,
+            Reviewing,
+            Completed,
+        }
+    }
+}
