+++ BackEnd/Profile/AutoMapperProfile.cs
@@ -1,4 +1,3 @@
-using AutoMapper;
 using BackEnd.Models;
 
 namespace BackEnd.Profile
